The Characters float and face in the right direction
Added one script to the project
Add the script to the character, do not make it a child object of the ball.
Drag and drop the ball into the public transform on the script
The character will float above this ball and look in the direction it is facing
Added a public distance above ball variable to change the height it floats above depending on the specific character so it looks good.